Move tests from the model over to the JUnit package.
