+++ src/JUnit/NonProfitContactTest.java
@@ -1,4 +1,4 @@
-package model;
+package JUnit;
 
 import static org.junit.Assert.*;
 
@@ -6,6 +6,9 @@ import java.time.LocalDate;
 
 import org.junit.Before;
 import org.junit.Test;
+
+import model.Auction;
+import model.NonProfitContact;
 
 /**
  *
